requirements.txt
fastapi==0.85.1
uvicorn[standard]==0.19.0
vulcan-api==2.2.1